- changed normalised image computation

+++ utils.py
@@ -359,7 +359,7 @@ def get_normalised_image(
     :param masked: Whether to mask the image or not
     :return:
     """
-    image = load_nii(image_name).get_fdata().astype(dtype)
+    image = load_nii(image_name).get_fdata()
 
     # If no mask is provided we use the image as a mask (all non-zero values)
     if mask is None:
@@ -373,7 +373,7 @@ def get_normalised_image(
             image_i = image[..., i]
             image_mu = np.mean(image_i[mask_bin])
             image_sigma = np.std(image_i[mask_bin])
-            norm_image = (image_i - image_mu) / image_sigma
+            norm_image = ((image_i - image_mu) / image_sigma).astype(dtype)
             if masked:
                 image_list.append(norm_image * mask_bin.astype(dtype))
             else:
@@ -384,7 +384,7 @@ def get_normalised_image(
         # Parameter estimation using the mask provided
         image_mu = np.mean(image[mask_bin])
         image_sigma = np.std(image[mask_bin])
-        norm_image = (image - image_mu) / image_sigma
+        norm_image = ((image - image_mu) / image_sigma).astype(dtype)
 
         if masked:
             output = norm_image * mask_bin.astype(dtype)
